pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-ai-agent"
version = "1.0.0"
description = "AI Agent with MCP support using Gemini SDK for intelligent system monitoring and automation"
authors = [{name = "MCP AI Agent Team"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-generativeai>=0.3.0",
    "mcp>=0.5.0",
    "asyncio-mqtt>=0.11.0",
    "aiofiles>=23.0.0",
    "pydantic>=2.0.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "psutil>=5.9.0",
    "watchdog>=3.0.0",
    "croniter>=1.4.0",
    "structlog>=23.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0"
]

[project.scripts]
mcp-agent = "mcp_ai_agent.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true